Adds uses to all Wagtail models.

+++ wagtail/admin/views/pages.py
@@ -28,6 +28,7 @@ from wagtail.admin.forms.search import SearchForm
 from wagtail.admin.mail import send_notification
 from wagtail.admin.navigation import get_explorable_root_page
 from wagtail.core import hooks
+from wagtail.core.collectors import get_paginated_uses
 from wagtail.core.models import Page, PageRevision, UserPagePermissionsProxy
 from wagtail.search.query import MATCH_ALL
 from wagtail.search.utils import parse_query_string
@@ -609,7 +610,9 @@ def delete(request, page_id):
 
         next_url = get_valid_next_url_from_request(request)
 
-        if request.method == 'POST':
+        uses = get_paginated_uses(request, page)
+
+        if request.method == 'POST' and not uses.are_protected:
             parent_id = page.get_parent().id
             page.delete()
 
@@ -626,6 +629,7 @@ def delete(request, page_id):
 
     return TemplateResponse(request, 'wagtailadmin/pages/confirm_delete.html', {
         'page': page,
+        'uses': uses,
         'descendant_count': page.get_descendant_count(),
         'next': next_url,
     })
