+++ wagtail/contrib/search_promotions/views.py
@@ -10,6 +10,7 @@ from wagtail.admin import messages
 from wagtail.admin.auth import any_permission_required, permission_required
 from wagtail.admin.forms.search import SearchForm
 from wagtail.contrib.search_promotions import forms
+from wagtail.core.collectors import get_paginated_uses
 from wagtail.search import forms as search_forms
 from wagtail.search.models import Query
 
@@ -166,11 +167,14 @@ def edit(request, query_id):
 def delete(request, query_id):
     query = get_object_or_404(Query, id=query_id)
 
-    if request.method == 'POST':
+    uses = get_paginated_uses(request, query)
+
+    if request.method == 'POST' and not uses.are_protected:
         query.editors_picks.all().delete()
         messages.success(request, _("Editor's picks deleted."))
         return redirect('wagtailsearchpromotions:index')
 
     return TemplateResponse(request, 'wagtailsearchpromotions/confirm_delete.html', {
         'query': query,
+        'uses': uses,
     })
