+++ wagtail/documents/views/documents.py
@@ -11,6 +11,7 @@ from wagtail.admin import messages
 from wagtail.admin.auth import PermissionPolicyChecker, permission_denied
 from wagtail.admin.forms.search import SearchForm
 from wagtail.admin.models import popular_tags_for_model
+from wagtail.core.collectors import get_paginated_uses
 from wagtail.core.models import Collection
 from wagtail.documents import get_document_model
 from wagtail.documents.forms import get_document_form
@@ -205,13 +206,16 @@ def delete(request, document_id):
     if not permission_policy.user_has_permission_for_instance(request.user, 'delete', doc):
         return permission_denied(request)
 
-    if request.method == 'POST':
+    uses = get_paginated_uses(request, doc)
+
+    if request.method == 'POST' and not uses.are_protected:
         doc.delete()
         messages.success(request, _("Document '{0}' deleted.").format(doc.title))
         return redirect('wagtaildocs:index')
 
     return TemplateResponse(request, "wagtaildocs/documents/confirm_delete.html", {
         'document': doc,
+        'uses': uses,
     })
