+++ wagtail/snippets/views/snippets.py
@@ -16,6 +16,7 @@ from wagtail.admin.auth import permission_denied
 from wagtail.admin.edit_handlers import ObjectList, extract_panel_definitions_from_model_class
 from wagtail.admin.forms.search import SearchForm
 from wagtail.core import hooks
+from wagtail.core.collectors import get_paginated_uses
 from wagtail.search.backends import get_search_backend
 from wagtail.search.index import class_is_indexed
 from wagtail.snippets.models import get_snippet_models
@@ -249,7 +250,9 @@ def delete(request, app_label, model_name, pk=None):
 
     count = len(instances)
 
-    if request.method == 'POST':
+    uses = get_paginated_uses(request, *instances)
+
+    if request.method == 'POST' and not uses.are_protected:
         for instance in instances:
             instance.delete()
 
@@ -288,6 +291,7 @@ def delete(request, app_label, model_name, pk=None):
             reverse('wagtailsnippets:delete-multiple', args=(app_label, model_name))
             + '?' + urlencode([('id', instance.pk) for instance in instances])
         ),
+        'uses': uses,
     })
